Simplified textUpdate widget No longer uses material-ui paper, simply uses typography object with emphasised background colour
