+++ src/malcolmWidgets/checkbox/WidgetCheckbox.component.js
