+++ src/malcolmWidgets/led/widgetLED.component.js
