+++ src/malcolmWidgets/textUpdate/WidgetTextUpdate.component.js
@@ -1,0 +1,35 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import { withStyles } from 'material-ui/styles';
+import Typography from 'material-ui/Typography';
+import { emphasize } from 'material-ui/styles/colorManipulator';
+
+const styles = theme => ({
+  textUpdate: {
+    backgroundColor: emphasize(theme.palette.background.paper, 0.1),
+    padding: 4,
+    textAlign: 'Right',
+    marginLeft: 4,
+    marginRight: 4,
+  },
+});
+
+const WidgetTextUpdate = props => (
+  <Typography className={props.classes.textUpdate}>
+    {props.Text} {props.Units ? props.Units : null}
+  </Typography>
+);
+
+WidgetTextUpdate.propTypes = {
+  Text: PropTypes.string.isRequired,
+  Units: PropTypes.string,
+  classes: PropTypes.shape({
+    textUpdate: PropTypes.string,
+  }).isRequired,
+};
+
+WidgetTextUpdate.defaultProps = {
+  Units: null,
+};
+
+export default withStyles(styles, { withTheme: true })(WidgetTextUpdate);
